requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
db-sqlite3
python-dateutil==2.8.2
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
mitosheet==0.1.567
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
plotly==5.18.0
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scipy==1.10.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.31.0
streamlit_option_menu==0.3.12
streamlit_authenticator==0.3.1
streamlit_pandas_profiling
sympy==1.12
st-paywall==0.1.2
tenacity==8.2.3
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.10
urllib3==2.2.0
validators==0.22.0
zipp==3.17.0
